requirements.txt
asyncpg
fastapi==0.101.0
psycopg2-binary==2.9.7
uvicorn==0.17.4
python-dotenv==0.15.0
black==22.3.0
isort==5.10.1
autoflake==1.4
requests==2.28.1
shortuuid==1.0.9
python-multipart==0.0.6
phonenumbers==8.13.18
pydantic==1.10.13
pydantic_core==2.10.1
pymongo==4.4.1
motor==3.2.0
pillow==10.0.0
